- 수정, 삭제, 추가 페이지에서 하단 테이블의 col을 병합 (colspan값을 임의로 100을 주었음)

+++ src/main/webapp/bookmark-group.jsp
@@ -39,20 +39,25 @@
         out.write("<th>수정일자</th>");
         out.write("<th>비고</th>");
         out.write("</tr>");
-        for (BookMarkGroup ele : bookMarkGroupList) {
-            out.write("<tr" + "id="+ele.getId() + ">");
-            out.write("<td>" + ele.getId() + "</td>");
-            out.write("<td>" + ele.getBookmarkGroupName() + "</td>");
-            out.write("<td>" + ele.getOrders() + "</td>");
-            out.write("<td>" + ele.getRegisterDate() + "</td>");
-            if (ele.getUpdateDate() != null) {
-                out.write("<td>" + ele.getUpdateDate() + "</td>");
-            } else {
-                out.write("<td></td>");
+        if (bookMarkGroupList.isEmpty()) {
+            out.write("<tr><td colspan=\"100\">등록된 북마크 그룹명이 없습니다.</td></tr>");
+        } else {
+            for (BookMarkGroup ele : bookMarkGroupList) {
+                out.write("<tr" + "id="+ele.getId() + ">");
+                out.write("<td>" + ele.getId() + "</td>");
+                out.write("<td>" + ele.getBookmarkGroupName() + "</td>");
+                out.write("<td>" + ele.getOrders() + "</td>");
+                out.write("<td>" + ele.getRegisterDate() + "</td>");
+                if (ele.getUpdateDate() != null) {
+                    out.write("<td>" + ele.getUpdateDate() + "</td>");
+                } else {
+                    out.write("<td></td>");
+                }
+                out.write(String.format("<td style=\"text-align:center;\"><a href=\"./bookmark-group-edit.jsp?id=%d\">수정</a> <a href=\"./bookmark-group-delete.jsp?id=%d\">삭제</a></td>", ele.getId(), ele.getId()));
+                out.write("</tr>");
             }
-            out.write(String.format("<td style=\"text-align:center;\"><a href=\"./bookmark-group-edit.jsp?id=%d\">수정</a> <a href=\"./bookmark-group-delete.jsp?id=%d\">삭제</a></td>", ele.getId(), ele.getId()));
-            out.write("</tr>");
         }
+
         out.write("</table>");
     %>
     <script>
